Move more datatable enums/flags to their headers
