+++ source/D2Common/include/D2DataTbls.h
@@ -163,6 +163,26 @@ struct D2StorePageTxt
 	uint32_t dwCode;						//0x00
 };
 
+enum D2C_PetTypeTxtFlagIndices
+{
+	PETTYPEFLAGINDEX_WARP = 0,
+	PETTYPEFLAGINDEX_RANGE,
+	PETTYPEFLAGINDEX_PARTYSEND,
+	PETTYPEFLAGINDEX_UNSUMMON,
+	PETTYPEFLAGINDEX_AUTOMAP,
+	PETTYPEFLAGINDEX_DRAWHP,
+};
+
+enum D2C_PetTypeTxtFlags
+{
+	PETTYPEFLAG_WARP = (1 << PETTYPEFLAGINDEX_WARP),
+	PETTYPEFLAG_RANGE = (1 << PETTYPEFLAGINDEX_RANGE),
+	PETTYPEFLAG_PARTYSEND = (1 << PETTYPEFLAGINDEX_PARTYSEND),
+	PETTYPEFLAG_UNSUMMON = (1 << PETTYPEFLAGINDEX_UNSUMMON),
+	PETTYPEFLAG_AUTOMAP = (1 << PETTYPEFLAGINDEX_AUTOMAP),
+	PETTYPEFLAG_DRAWHP = (1 << PETTYPEFLAGINDEX_DRAWHP),
+};
+
 struct D2PetTypeTxt
 {
 	uint32_t dwHcIdx;						//0x00
