+++ source/D2Common/include/D2States.h
@@ -2,6 +2,49 @@
 
 #include "CommonDefinitions.h"
 
+enum D2C_StateMasks
+{
+	STATEMASK_NOSEND,
+	STATEMASK_AURA,
+	STATEMASK_HIDE,
+	STATEMASK_TRANSFORM,
+	STATEMASK_PGSV,
+	STATEMASK_ACTIVE,
+	STATEMASK_REMHIT,
+	STATEMASK_DAMBLUE,
+	STATEMASK_DAMRED,
+	STATEMASK_ATTBLUE,
+	STATEMASK_ATTRED,
+	STATEMASK_CURSE,
+	STATEMASK_CURABLE,
+	STATEMASK_PLRSTAYDEATH,
+	STATEMASK_MONSTAYDEATH,
+	STATEMASK_BOSSSTAYDEATH,
+	STATEMASK_DISGUISE,
+	STATEMASK_RESTRICT,
+	STATEMASK_BLUE,
+	STATEMASK_ARMBLUE,
+	STATEMASK_RFBLUE,
+	STATEMASK_RCBLUE,
+	STATEMASK_RLBLUE,
+	STATEMASK_RPBLUE,
+	STATEMASK_STAMBARBLUE,
+	STATEMASK_ARMRED,
+	STATEMASK_RFRED,
+	STATEMASK_RCRED,
+	STATEMASK_RLRED,
+	STATEMASK_RPRED,
+	STATEMASK_EXP,
+	STATEMASK_SHATTER,
+	STATEMASK_LIFE,
+	STATEMASK_UDEAD,
+	STATEMASK_GREEN,
+	STATEMASK_NOOVERLAYS,
+	STATEMASK_NOCLEAR,
+	STATEMASK_BOSSINV,
+	STATEMASK_MELEEONLY,
+	STATEMASK_NOTONDEAD,
+};
 
 //D2Common.0x6FDB4560 (#10486)
 void __stdcall STATES_ToggleState(D2UnitStrc* pUnit, int nState, BOOL bSet);
