+++ source/D2Common/include/DataTbls/ItemsTbls.h
@@ -100,6 +100,22 @@ struct D2SetsTxt
 	D2SetItemsTxt* pSetItem[6];				//0x110
 };
 
+enum D2C_UniqueItemsTxtFlagIndices
+{
+	UNIQUEITEMSFLAGINDEX_ENABLED = 0,
+	UNIQUEITEMSFLAGINDEX_NOLIMIT,
+	UNIQUEITEMSFLAGINDEX_CARRY1,
+	UNIQUEITEMSFLAGINDEX_LADDER,
+};
+
+enum D2C_UniqueItemsTxtFlags
+{
+	UNIQUEITEMSFLAG_ENABLED = (1 << UNIQUEITEMSFLAGINDEX_ENABLED),
+	UNIQUEITEMSFLAG_NOLIMIT = (1 << UNIQUEITEMSFLAGINDEX_NOLIMIT),
+	UNIQUEITEMSFLAG_CARRY1 = (1 << UNIQUEITEMSFLAGINDEX_CARRY1),
+	UNIQUEITEMSFLAG_LADDER = (1 << UNIQUEITEMSFLAGINDEX_LADDER),
+};
+
 struct D2UniqueItemsTxt
 {
 	uint16_t wId;							//0x00
@@ -163,6 +179,46 @@ struct D2OpStatDataStrc
 	uint16_t nStat;							//0x02
 	uint8_t nOp;							//0x04
 	uint8_t nOpParam;						//0x05
+};
+
+enum D2C_ItemStatCostTxtFlagIndices
+{
+	ITEMSTATCOSTFLAGINDEX_SEND_OTHER = 0,
+	ITEMSTATCOSTFLAGINDEX_SIGNED,
+	ITEMSTATCOSTFLAGINDEX_DAMAGERELATED,
+	ITEMSTATCOSTFLAGINDEX_ITEMSPECIFIC,
+	ITEMSTATCOSTFLAGINDEX_DIRECT,
+
+	ITEMSTATCOSTFLAGINDEX_HP_MANA_STAMINA = 5,
+	ITEMSTATCOSTFLAGINDEX_HP,
+	ITEMSTATCOSTFLAGINDEX_MANA,
+	ITEMSTATCOSTFLAGINDEX_STAMINA,
+
+	ITEMSTATCOSTFLAGINDEX_UPDATEANIMRATE = 9,
+	ITEMSTATCOSTFLAGINDEX_FMIN,
+	ITEMSTATCOSTFLAGINDEX_FCALLBACK,
+	ITEMSTATCOSTFLAGINDEX_SAVED,
+	ITEMSTATCOSTFLAGINDEX_CSVSIGNED,
+};
+
+enum D2C_ItemStatCostTxtFlags
+{
+	ITEMSTATCOSTFLAG_SEND_OTHER = (1 << ITEMSTATCOSTFLAGINDEX_SEND_OTHER),
+	ITEMSTATCOSTFLAG_SIGNED = (1 << ITEMSTATCOSTFLAGINDEX_SIGNED),
+	ITEMSTATCOSTFLAG_DAMAGERELATED = (1 << ITEMSTATCOSTFLAGINDEX_DAMAGERELATED),
+	ITEMSTATCOSTFLAG_ITEMSPECIFIC = (1 << ITEMSTATCOSTFLAGINDEX_ITEMSPECIFIC),
+	ITEMSTATCOSTFLAG_DIRECT = (1 << ITEMSTATCOSTFLAGINDEX_DIRECT),
+
+	ITEMSTATCOSTFLAG_HP_MANA_STAMINA = (1 << ITEMSTATCOSTFLAGINDEX_HP_MANA_STAMINA),
+	ITEMSTATCOSTFLAG_HP = (1 << ITEMSTATCOSTFLAGINDEX_HP),
+	ITEMSTATCOSTFLAG_MANA = (1 << ITEMSTATCOSTFLAGINDEX_MANA),
+	ITEMSTATCOSTFLAG_STAMINA = (1 << ITEMSTATCOSTFLAGINDEX_STAMINA),
+
+	ITEMSTATCOSTFLAG_UPDATEANIMRATE = (1 << ITEMSTATCOSTFLAGINDEX_UPDATEANIMRATE),
+	ITEMSTATCOSTFLAG_FMIN = (1 << ITEMSTATCOSTFLAGINDEX_FMIN),
+	ITEMSTATCOSTFLAG_FCALLBACK = (1 << ITEMSTATCOSTFLAGINDEX_FCALLBACK),
+	ITEMSTATCOSTFLAG_SAVED = (1 << ITEMSTATCOSTFLAGINDEX_SAVED),
+	ITEMSTATCOSTFLAG_CSVSIGNED = (1 << ITEMSTATCOSTFLAGINDEX_CSVSIGNED),
 };
 
 struct D2ItemStatCostTxt
