+++ source/D2Common/include/DataTbls/MissilesTbls.h
@@ -2,6 +2,46 @@
 
 #include <D2BasicTypes.h>
 
+enum D2C_MissilesTxtFlagIndices
+{
+	MISSILESFLAGINDEX_LASTCOLLIDE = 0,
+	MISSILESFLAGINDEX_EXPLOSION,
+	MISSILESFLAGINDEX_PIERCE,
+	MISSILESFLAGINDEX_CANSLOW,
+	MISSILESFLAGINDEX_CANDESTROY,
+	MISSILESFLAGINDEX_CLIENTSEND,
+	MISSILESFLAGINDEX_GETHIT,
+	MISSILESFLAGINDEX_SOFTHIT,
+	MISSILESFLAGINDEX_APPLYMASTERY,
+	MISSILESFLAGINDEX_RETURNFIRE,
+	MISSILESFLAGINDEX_TOWN,
+	MISSILESFLAGINDEX_SRCTOWN,
+	MISSILESFLAGINDEX_NOMULTISHOT,
+	MISSILESFLAGINDEX_NOUNIQUEMOD,
+	MISSILESFLAGINDEX_HALF2HSRC,
+	MISSILESFLAGINDEX_MISSILESKILL,
+};
+
+enum D2C_MissilesTxtFlags
+{
+	MISSILESFLAG_LASTCOLLIDE = (1 << MISSILESFLAGINDEX_LASTCOLLIDE),
+	MISSILESFLAG_EXPLOSION = (1 << MISSILESFLAGINDEX_EXPLOSION),
+	MISSILESFLAG_PIERCE = (1 << MISSILESFLAGINDEX_PIERCE),
+	MISSILESFLAG_CANSLOW = (1 << MISSILESFLAGINDEX_CANSLOW),
+	MISSILESFLAG_CANDESTROY = (1 << MISSILESFLAGINDEX_CANDESTROY),
+	MISSILESFLAG_CLIENTSEND = (1 << MISSILESFLAGINDEX_CLIENTSEND),
+	MISSILESFLAG_GETHIT = (1 << MISSILESFLAGINDEX_GETHIT),
+	MISSILESFLAG_SOFTHIT = (1 << MISSILESFLAGINDEX_SOFTHIT),
+	MISSILESFLAG_APPLYMASTERY = (1 << MISSILESFLAGINDEX_APPLYMASTERY),
+	MISSILESFLAG_RETURNFIRE = (1 << MISSILESFLAGINDEX_RETURNFIRE),
+	MISSILESFLAG_TOWN = (1 << MISSILESFLAGINDEX_TOWN),
+	MISSILESFLAG_SRCTOWN = (1 << MISSILESFLAGINDEX_SRCTOWN),
+	MISSILESFLAG_NOMULTISHOT = (1 << MISSILESFLAGINDEX_NOMULTISHOT),
+	MISSILESFLAG_NOUNIQUEMOD = (1 << MISSILESFLAGINDEX_NOUNIQUEMOD),
+	MISSILESFLAG_HALF2HSRC = (1 << MISSILESFLAGINDEX_HALF2HSRC),
+	MISSILESFLAG_MISSILESKILL = (1 << MISSILESFLAGINDEX_MISSILESKILL),
+};
+
 struct D2MissilesTxt
 {
 	uint16_t wId;							//0x00
