+++ source/D2Common/include/DataTbls/MonsterTbls.h
@@ -148,6 +148,202 @@ struct D2MonSoundsTxt
 	uint32_t dwCvtSk3;						//0x90
 };
 
+enum D2C_MonStatsTxtFlagIndices
+{
+	MONSTATSFLAGINDEX_ISSPAWN = 0,
+	MONSTATSFLAGINDEX_ISMELEE,
+	MONSTATSFLAGINDEX_NORATIO,
+	MONSTATSFLAGINDEX_OPENDOORS,
+	MONSTATSFLAGINDEX_SETBOSS,
+	MONSTATSFLAGINDEX_BOSSXFER,
+	MONSTATSFLAGINDEX_BOSS,
+	MONSTATSFLAGINDEX_PRIMEEVIL,
+	MONSTATSFLAGINDEX_NPC,
+	MONSTATSFLAGINDEX_INTERACT,
+	MONSTATSFLAGINDEX_INTOWN,
+	MONSTATSFLAGINDEX_LUNDEAD,
+	MONSTATSFLAGINDEX_HUNDEAD,
+	MONSTATSFLAGINDEX_DEMON,
+	MONSTATSFLAGINDEX_FLYING,
+	MONSTATSFLAGINDEX_KILLABLE,
+	MONSTATSFLAGINDEX_SWITCHAI,
+	MONSTATSFLAGINDEX_NOMULTISHOT,
+	MONSTATSFLAGINDEX_NEVERCOUNT,
+	MONSTATSFLAGINDEX_PETIGNORE,
+	MONSTATSFLAGINDEX_DEATHDMG,
+	MONSTATSFLAGINDEX_GENERICSPAWN,
+	MONSTATSFLAGINDEX_ZOO,
+	MONSTATSFLAGINDEX_PLACESPAWN,
+	MONSTATSFLAGINDEX_INVENTORY,
+	MONSTATSFLAGINDEX_ENABLED,
+	MONSTATSFLAGINDEX_NOSHLDBLOCK,
+	MONSTATSFLAGINDEX_NOAURA,
+	MONSTATSFLAGINDEX_RANGEDTYPE,
+};
+
+enum D2C_MonStatsTxtFlags
+{
+	MONSTATSFLAG_ISSPAWN = (1 << MONSTATSFLAGINDEX_ISSPAWN),
+	MONSTATSFLAG_ISMELEE = (1 << MONSTATSFLAGINDEX_ISMELEE),
+	MONSTATSFLAG_NORATIO = (1 << MONSTATSFLAGINDEX_NORATIO),
+	MONSTATSFLAG_OPENDOORS = (1 << MONSTATSFLAGINDEX_OPENDOORS),
+	MONSTATSFLAG_SETBOSS = (1 << MONSTATSFLAGINDEX_SETBOSS),
+	MONSTATSFLAG_BOSSXFER = (1 << MONSTATSFLAGINDEX_BOSSXFER),
+	MONSTATSFLAG_BOSS = (1 << MONSTATSFLAGINDEX_BOSS),
+	MONSTATSFLAG_PRIMEEVIL = (1 << MONSTATSFLAGINDEX_PRIMEEVIL),
+	MONSTATSFLAG_NPC = (1 << MONSTATSFLAGINDEX_NPC),
+	MONSTATSFLAG_INTERACT = (1 << MONSTATSFLAGINDEX_INTERACT),
+	MONSTATSFLAG_INTOWN = (1 << MONSTATSFLAGINDEX_INTOWN),
+	MONSTATSFLAG_LUNDEAD = (1 << MONSTATSFLAGINDEX_LUNDEAD),
+	MONSTATSFLAG_HUNDEAD = (1 << MONSTATSFLAGINDEX_HUNDEAD),
+	MONSTATSFLAG_DEMON = (1 << MONSTATSFLAGINDEX_DEMON),
+	MONSTATSFLAG_FLYING = (1 << MONSTATSFLAGINDEX_FLYING),
+	MONSTATSFLAG_KILLABLE = (1 << MONSTATSFLAGINDEX_KILLABLE),
+	MONSTATSFLAG_SWITCHAI = (1 << MONSTATSFLAGINDEX_SWITCHAI),
+	MONSTATSFLAG_NOMULTISHOT = (1 << MONSTATSFLAGINDEX_NOMULTISHOT),
+	MONSTATSFLAG_NEVERCOUNT = (1 << MONSTATSFLAGINDEX_NEVERCOUNT),
+	MONSTATSFLAG_PETIGNORE = (1 << MONSTATSFLAGINDEX_PETIGNORE),
+	MONSTATSFLAG_DEATHDMG = (1 << MONSTATSFLAGINDEX_DEATHDMG),
+	MONSTATSFLAG_GENERICSPAWN = (1 << MONSTATSFLAGINDEX_GENERICSPAWN),
+	MONSTATSFLAG_ZOO = (1 << MONSTATSFLAGINDEX_ZOO),
+	MONSTATSFLAG_PLACESPAWN = (1 << MONSTATSFLAGINDEX_PLACESPAWN),
+	MONSTATSFLAG_INVENTORY = (1 << MONSTATSFLAGINDEX_INVENTORY),
+	MONSTATSFLAG_ENABLED = (1 << MONSTATSFLAGINDEX_ENABLED),
+	MONSTATSFLAG_NOSHLDBLOCK = (1 << MONSTATSFLAGINDEX_NOSHLDBLOCK),
+	MONSTATSFLAG_NOAURA = (1 << MONSTATSFLAGINDEX_NOAURA),
+	MONSTATSFLAG_RANGEDTYPE = (1 << MONSTATSFLAGINDEX_RANGEDTYPE),
+};
+
+enum D2C_MonStats2TxtFlagIndices
+{
+	MONSTATS2FLAGINDEX_NOGFXHITTEST = 0,
+	MONSTATS2FLAGINDEX_NOMAP,
+	MONSTATS2FLAGINDEX_NOOVLY,
+	MONSTATS2FLAGINDEX_ISSEL,
+	MONSTATS2FLAGINDEX_ALSEL,
+	MONSTATS2FLAGINDEX_NOSEL,
+	MONSTATS2FLAGINDEX_SHIFTSEL,
+	MONSTATS2FLAGINDEX_CORPSESEL,
+	MONSTATS2FLAGINDEX_REVIVE,
+	MONSTATS2FLAGINDEX_ISATT,
+	MONSTATS2FLAGINDEX_SMALL,
+	MONSTATS2FLAGINDEX_LARGE,
+	MONSTATS2FLAGINDEX_SOFT,
+	MONSTATS2FLAGINDEX_CRITTER,
+	MONSTATS2FLAGINDEX_SHADOW,
+	MONSTATS2FLAGINDEX_NOUNIQUESHIFT,
+	MONSTATS2FLAGINDEX_COMPOSITEDEATH,
+	MONSTATS2FLAGINDEX_INERT,
+	MONSTATS2FLAGINDEX_OBJCOL,
+	MONSTATS2FLAGINDEX_DEADCOL,
+	MONSTATS2FLAGINDEX_UNFLATDEAD,
+};
+
+enum D2C_MonStats2TxtFlags
+{
+	MONSTATS2FLAG_NOGFXHITTEST = (1 << MONSTATS2FLAGINDEX_NOGFXHITTEST),
+	MONSTATS2FLAG_NOMAP = (1 << MONSTATS2FLAGINDEX_NOMAP),
+	MONSTATS2FLAG_NOOVLY = (1 << MONSTATS2FLAGINDEX_NOOVLY),
+	MONSTATS2FLAG_ISSEL = (1 << MONSTATS2FLAGINDEX_ISSEL),
+	MONSTATS2FLAG_ALSEL = (1 << MONSTATS2FLAGINDEX_ALSEL),
+	MONSTATS2FLAG_NOSEL = (1 << MONSTATS2FLAGINDEX_NOSEL),
+	MONSTATS2FLAG_SHIFTSEL = (1 << MONSTATS2FLAGINDEX_SHIFTSEL),
+	MONSTATS2FLAG_CORPSESEL = (1 << MONSTATS2FLAGINDEX_CORPSESEL),
+	MONSTATS2FLAG_REVIVE = (1 << MONSTATS2FLAGINDEX_REVIVE),
+	MONSTATS2FLAG_ISATT = (1 << MONSTATS2FLAGINDEX_ISATT),
+	MONSTATS2FLAG_SMALL = (1 << MONSTATS2FLAGINDEX_SMALL),
+	MONSTATS2FLAG_LARGE = (1 << MONSTATS2FLAGINDEX_LARGE),
+	MONSTATS2FLAG_SOFT = (1 << MONSTATS2FLAGINDEX_SOFT),
+	MONSTATS2FLAG_CRITTER = (1 << MONSTATS2FLAGINDEX_CRITTER),
+	MONSTATS2FLAG_SHADOW = (1 << MONSTATS2FLAGINDEX_SHADOW),
+	MONSTATS2FLAG_NOUNIQUESHIFT = (1 << MONSTATS2FLAGINDEX_NOUNIQUESHIFT),
+	MONSTATS2FLAG_COMPOSITEDEATH = (1 << MONSTATS2FLAGINDEX_COMPOSITEDEATH),
+	MONSTATS2FLAG_INERT = (1 << MONSTATS2FLAGINDEX_INERT),
+	MONSTATS2FLAG_OBJCOL = (1 << MONSTATS2FLAGINDEX_OBJCOL),
+	MONSTATS2FLAG_DEADCOL = (1 << MONSTATS2FLAGINDEX_DEADCOL),
+	MONSTATS2FLAG_UNFLATDEAD = (1 << MONSTATS2FLAGINDEX_UNFLATDEAD),
+};
+
+enum D2C_MonStats2TxtComponentFlagIndices
+{
+	MONSTATS2COMPONENTFLAGINDEX_HD = 0,
+	MONSTATS2COMPONENTFLAGINDEX_TR,
+	MONSTATS2COMPONENTFLAGINDEX_LG,
+	MONSTATS2COMPONENTFLAGINDEX_RA,
+	MONSTATS2COMPONENTFLAGINDEX_LA,
+	MONSTATS2COMPONENTFLAGINDEX_RH,
+	MONSTATS2COMPONENTFLAGINDEX_LH,
+	MONSTATS2COMPONENTFLAGINDEX_SH,
+	MONSTATS2COMPONENTFLAGINDEX_S1,
+	MONSTATS2COMPONENTFLAGINDEX_S2,
+	MONSTATS2COMPONENTFLAGINDEX_S3,
+	MONSTATS2COMPONENTFLAGINDEX_S4,
+	MONSTATS2COMPONENTFLAGINDEX_S5,
+	MONSTATS2COMPONENTFLAGINDEX_S6,
+	MONSTATS2COMPONENTFLAGINDEX_S7,
+	MONSTATS2COMPONENTFLAGINDEX_S8,
+};
+
+enum D2C_MonStats2TxtComponentFlags
+{
+	MONSTATS2COMPONENTFLAG_HD = (1 << MONSTATS2COMPONENTFLAGINDEX_HD),
+	MONSTATS2COMPONENTFLAG_TR = (1 << MONSTATS2COMPONENTFLAGINDEX_TR),
+	MONSTATS2COMPONENTFLAG_LG = (1 << MONSTATS2COMPONENTFLAGINDEX_LG),
+	MONSTATS2COMPONENTFLAG_RA = (1 << MONSTATS2COMPONENTFLAGINDEX_RA),
+	MONSTATS2COMPONENTFLAG_LA = (1 << MONSTATS2COMPONENTFLAGINDEX_LA),
+	MONSTATS2COMPONENTFLAG_RH = (1 << MONSTATS2COMPONENTFLAGINDEX_RH),
+	MONSTATS2COMPONENTFLAG_LH = (1 << MONSTATS2COMPONENTFLAGINDEX_LH),
+	MONSTATS2COMPONENTFLAG_SH = (1 << MONSTATS2COMPONENTFLAGINDEX_SH),
+	MONSTATS2COMPONENTFLAG_S1 = (1 << MONSTATS2COMPONENTFLAGINDEX_S1),
+	MONSTATS2COMPONENTFLAG_S2 = (1 << MONSTATS2COMPONENTFLAGINDEX_S2),
+	MONSTATS2COMPONENTFLAG_S3 = (1 << MONSTATS2COMPONENTFLAGINDEX_S3),
+	MONSTATS2COMPONENTFLAG_S4 = (1 << MONSTATS2COMPONENTFLAGINDEX_S4),
+	MONSTATS2COMPONENTFLAG_S5 = (1 << MONSTATS2COMPONENTFLAGINDEX_S5),
+	MONSTATS2COMPONENTFLAG_S6 = (1 << MONSTATS2COMPONENTFLAGINDEX_S6),
+	MONSTATS2COMPONENTFLAG_S7 = (1 << MONSTATS2COMPONENTFLAGINDEX_S7),
+	MONSTATS2COMPONENTFLAG_S8 = (1 << MONSTATS2COMPONENTFLAGINDEX_S8),
+};
+
+enum D2C_MonStats2TxtModeFlagIndices
+{
+	MONSTATS2MODEFLAGINDEX_DT = 0,
+	MONSTATS2MODEFLAGINDEX_NU,
+	MONSTATS2MODEFLAGINDEX_WL,
+	MONSTATS2MODEFLAGINDEX_GH,
+	MONSTATS2MODEFLAGINDEX_A1,
+	MONSTATS2MODEFLAGINDEX_A2,
+	MONSTATS2MODEFLAGINDEX_BL,
+	MONSTATS2MODEFLAGINDEX_SC,
+	MONSTATS2MODEFLAGINDEX_S1,
+	MONSTATS2MODEFLAGINDEX_S2,
+	MONSTATS2MODEFLAGINDEX_S3,
+	MONSTATS2MODEFLAGINDEX_S4,
+	MONSTATS2MODEFLAGINDEX_DD,
+	MONSTATS2MODEFLAGINDEX_KB,
+	MONSTATS2MODEFLAGINDEX_SQ,
+	MONSTATS2MODEFLAGINDEX_RN,
+};
+
+enum D2C_MonStats2TxtModeFlags
+{
+	MONSTATS2MODEFLAG_DT = (1 << MONSTATS2MODEFLAGINDEX_DT),
+	MONSTATS2MODEFLAG_NU = (1 << MONSTATS2MODEFLAGINDEX_NU),
+	MONSTATS2MODEFLAG_WL = (1 << MONSTATS2MODEFLAGINDEX_WL),
+	MONSTATS2MODEFLAG_GH = (1 << MONSTATS2MODEFLAGINDEX_GH),
+	MONSTATS2MODEFLAG_A1 = (1 << MONSTATS2MODEFLAGINDEX_A1),
+	MONSTATS2MODEFLAG_A2 = (1 << MONSTATS2MODEFLAGINDEX_A2),
+	MONSTATS2MODEFLAG_BL = (1 << MONSTATS2MODEFLAGINDEX_BL),
+	MONSTATS2MODEFLAG_SC = (1 << MONSTATS2MODEFLAGINDEX_SC),
+	MONSTATS2MODEFLAG_S1 = (1 << MONSTATS2MODEFLAGINDEX_S1),
+	MONSTATS2MODEFLAG_S2 = (1 << MONSTATS2MODEFLAGINDEX_S2),
+	MONSTATS2MODEFLAG_S3 = (1 << MONSTATS2MODEFLAGINDEX_S3),
+	MONSTATS2MODEFLAG_S4 = (1 << MONSTATS2MODEFLAGINDEX_S4),
+	MONSTATS2MODEFLAG_DD = (1 << MONSTATS2MODEFLAGINDEX_DD),
+	MONSTATS2MODEFLAG_KB = (1 << MONSTATS2MODEFLAGINDEX_KB),
+	MONSTATS2MODEFLAG_SQ = (1 << MONSTATS2MODEFLAGINDEX_SQ),
+	MONSTATS2MODEFLAG_RN = (1 << MONSTATS2MODEFLAGINDEX_RN),
+};
+
 struct D2MonStatsTxt
 {
 	int16_t nId;							//0x00
