+++ source/D2Common/include/DataTbls/SkillsTbls.h
@@ -3,6 +3,96 @@
 #include <D2BasicTypes.h>
 
 
+enum D2C_SkillsTxtFlagIndices
+{
+	SKILLSFLAGINDEX_DECQUANT = 0,
+	SKILLSFLAGINDEX_LOB,
+	SKILLSFLAGINDEX_PROGRESSIVE,
+	SKILLSFLAGINDEX_FINISHING,
+	SKILLSFLAGINDEX_PASSIVE,
+	SKILLSFLAGINDEX_AURA,
+	SKILLSFLAGINDEX_PERIODIC,
+	SKILLSFLAGINDEX_PRGSTACK,
+	SKILLSFLAGINDEX_INTOWN,
+	SKILLSFLAGINDEX_KICK,
+	SKILLSFLAGINDEX_INGAME,
+	SKILLSFLAGINDEX_REPEAT,
+	SKILLSFLAGINDEX_STSUCCESSONLY,
+	SKILLSFLAGINDEX_STSOUNDDELAY,
+	SKILLSFLAGINDEX_WEAPONSND,
+	SKILLSFLAGINDEX_IMMEDIATE,
+	SKILLSFLAGINDEX_NOAMMO,
+	SKILLSFLAGINDEX_ENHANCEABLE,
+	SKILLSFLAGINDEX_DURABILITY,
+	SKILLSFLAGINDEX_USEATTACKRATE,
+	SKILLSFLAGINDEX_TARGETABLEONLY,
+	SKILLSFLAGINDEX_SEARCHENEMYXY,
+	SKILLSFLAGINDEX_SEARCHENEMYNEAR,
+	SKILLSFLAGINDEX_SEARCHOPENXY,
+	SKILLSFLAGINDEX_TARGETCORPSE,
+	SKILLSFLAGINDEX_TARGETPET,
+	SKILLSFLAGINDEX_TARGETALLY,
+	SKILLSFLAGINDEX_TARGETITEM,
+	SKILLSFLAGINDEX_ATTACKNOMANA,
+	SKILLSFLAGINDEX_ITEMTGTDO,
+	SKILLSFLAGINDEX_LEFTSKILL,
+	SKILLSFLAGINDEX_INTERRUPT,
+
+	//Bits above 31 need to restart at 0
+	SKILLSFLAGINDEX2_TGTPLACECHECK = 0,
+	SKILLSFLAGINDEX2_ITEMCHECKSTART,
+	SKILLSFLAGINDEX2_ITEMCLTCHECKSTART,
+	SKILLSFLAGINDEX2_GENERAL,
+	SKILLSFLAGINDEX2_SCROLL,
+	SKILLSFLAGINDEX2_USEMANAONDO,
+	SKILLSFLAGINDEX2_WARP,
+};
+
+enum D2C_SkillsTxtFlags
+{
+	SKILLSFLAG_DECQUANT = (1 << SKILLSFLAGINDEX_DECQUANT),
+	SKILLSFLAG_LOB = (1 << SKILLSFLAGINDEX_LOB),
+	SKILLSFLAG_PROGRESSIVE = (1 << SKILLSFLAGINDEX_PROGRESSIVE),
+	SKILLSFLAG_FINISHING = (1 << SKILLSFLAGINDEX_FINISHING),
+	SKILLSFLAG_PASSIVE = (1 << SKILLSFLAGINDEX_PASSIVE),
+	SKILLSFLAG_AURA = (1 << SKILLSFLAGINDEX_AURA),
+	SKILLSFLAG_PERIODIC = (1 << SKILLSFLAGINDEX_PERIODIC),
+	SKILLSFLAG_PRGSTACK = (1 << SKILLSFLAGINDEX_PRGSTACK),
+	SKILLSFLAG_INTOWN = (1 << SKILLSFLAGINDEX_INTOWN),
+	SKILLSFLAG_KICK = (1 << SKILLSFLAGINDEX_KICK),
+	SKILLSFLAG_INGAME = (1 << SKILLSFLAGINDEX_INGAME),
+	SKILLSFLAG_REPEAT = (1 << SKILLSFLAGINDEX_REPEAT),
+	SKILLSFLAG_STSUCCESSONLY = (1 << SKILLSFLAGINDEX_STSUCCESSONLY),
+	SKILLSFLAG_STSOUNDDELAY = (1 << SKILLSFLAGINDEX_STSOUNDDELAY),
+	SKILLSFLAG_WEAPONSND = (1 << SKILLSFLAGINDEX_WEAPONSND),
+	SKILLSFLAG_IMMEDIATE = (1 << SKILLSFLAGINDEX_IMMEDIATE),
+	SKILLSFLAG_NOAMMO = (1 << SKILLSFLAGINDEX_NOAMMO),
+	SKILLSFLAG_ENHANCEABLE = (1 << SKILLSFLAGINDEX_ENHANCEABLE),
+	SKILLSFLAG_DURABILITY = (1 << SKILLSFLAGINDEX_DURABILITY),
+	SKILLSFLAG_USEATTACKRATE = (1 << SKILLSFLAGINDEX_USEATTACKRATE),
+	SKILLSFLAG_TARGETABLEONLY = (1 << SKILLSFLAGINDEX_TARGETABLEONLY),
+	SKILLSFLAG_SEARCHENEMYXY = (1 << SKILLSFLAGINDEX_SEARCHENEMYXY),
+	SKILLSFLAG_SEARCHENEMYNEAR = (1 << SKILLSFLAGINDEX_SEARCHENEMYNEAR),
+	SKILLSFLAG_SEARCHOPENXY = (1 << SKILLSFLAGINDEX_SEARCHOPENXY),
+	SKILLSFLAG_TARGETCORPSE = (1 << SKILLSFLAGINDEX_TARGETCORPSE),
+	SKILLSFLAG_TARGETPET = (1 << SKILLSFLAGINDEX_TARGETPET),
+	SKILLSFLAG_TARGETALLY = (1 << SKILLSFLAGINDEX_TARGETALLY),
+	SKILLSFLAG_TARGETITEM = (1 << SKILLSFLAGINDEX_TARGETITEM),
+	SKILLSFLAG_ATTACKNOMANA = (1 << SKILLSFLAGINDEX_ATTACKNOMANA),
+	SKILLSFLAG_ITEMTGTDO = (1 << SKILLSFLAGINDEX_ITEMTGTDO),
+	SKILLSFLAG_LEFTSKILL = (1 << SKILLSFLAGINDEX_LEFTSKILL),
+	SKILLSFLAG_INTERRUPT = (1 << SKILLSFLAGINDEX_INTERRUPT),
+
+	//Bits above 31 need to restart at 0
+	SKILLSFLAG2_TGTPLACECHECK = (1 << SKILLSFLAGINDEX2_TGTPLACECHECK),
+	SKILLSFLAG2_ITEMCHECKSTART = (1 << SKILLSFLAGINDEX2_ITEMCHECKSTART),
+	SKILLSFLAG2_ITEMCLTCHECKSTART = (1 << SKILLSFLAGINDEX2_ITEMCLTCHECKSTART),
+	SKILLSFLAG2_GENERAL = (1 << SKILLSFLAGINDEX2_GENERAL),
+	SKILLSFLAG2_SCROLL = (1 << SKILLSFLAGINDEX2_SCROLL),
+	SKILLSFLAG2_USEMANAONDO = (1 << SKILLSFLAGINDEX2_USEMANAONDO),
+	SKILLSFLAG2_WARP = (1 << SKILLSFLAGINDEX2_WARP),
+};
+
 struct D2SkillCalcTxt
 {
 	uint32_t dwCode;						//0x00
