+++ source/D2CommonDefinitions/include/D2Constants.h
@@ -2247,118 +2247,6 @@ enum D2C_OriginalQuestFlags
 //	QSTATE_QUESTSTARTEDNOTCOMPLETED,
 //};
 
-enum D2QuestStateFlagIds
-{
-	QUESTSTATEFLAG_A1Q0 = 0,
-	QUESTSTATEFLAG_A1Q1 = 1,
-	QUESTSTATEFLAG_A1Q2 = 2,
-	QUESTSTATEFLAG_A1Q3 = 3,
-	QUESTSTATEFLAG_A1Q4 = 4,
-	QUESTSTATEFLAG_A1Q5 = 5,
-	QUESTSTATEFLAG_A1Q6 = 6,
-
-	QUESTSTATEFLAG_A1COMPLETED = 7,
-
-	QUESTSTATEFLAG_A2Q0 = 8,
-	QUESTSTATEFLAG_A2Q1 = 9,
-	QUESTSTATEFLAG_A2Q2 = 10,
-	QUESTSTATEFLAG_A2Q3 = 11,
-	QUESTSTATEFLAG_A2Q4 = 12,
-	QUESTSTATEFLAG_A2Q5 = 13,
-	QUESTSTATEFLAG_A2Q6 = 14,
-
-	QUESTSTATEFLAG_A2COMPLETED = 15,
-
-	QUESTSTATEFLAG_A3Q0 = 16,
-	QUESTSTATEFLAG_A3Q1 = 17,
-	QUESTSTATEFLAG_A3Q2 = 18,
-	QUESTSTATEFLAG_A3Q3 = 19,
-	QUESTSTATEFLAG_A3Q4 = 20,
-	QUESTSTATEFLAG_A3Q5 = 21,
-	QUESTSTATEFLAG_A3Q6 = 22,
-
-	QUESTSTATEFLAG_A3COMPLETED = 23,
-
-	QUESTSTATEFLAG_A4Q0 = 24,
-	QUESTSTATEFLAG_A4Q1 = 25,
-	QUESTSTATEFLAG_A4Q2 = 26,
-	QUESTSTATEFLAG_A4Q3 = 27,
-
-	QUESTSTATEFLAG_A4COMPLETED = 28,
-
-	QUESTSTATEFLAG_A1Q7 = 29,
-
-	QUESTSTATEFLAG_A2Q7 = 30,
-	QUESTSTATEFLAG_A2Q8 = 31,
-
-	QUESTSTATEFLAG_A3Q7 = 32,
-
-	QUESTSTATEFLAG_A4Q4 = 33,
-
-	//Unused: 34
-
-	QUESTSTATEFLAG_A5Q1 = 35,
-	QUESTSTATEFLAG_A5Q2 = 36,
-	QUESTSTATEFLAG_A5Q3 = 37,
-	QUESTSTATEFLAG_A5Q4 = 38,
-	QUESTSTATEFLAG_A5Q5 = 39,
-	QUESTSTATEFLAG_A5Q6 = 40,
-
-	QUESTSTATEFLAG_A1INTRO = 41,
-	QUESTSTATEFLAG_A2INTRO = 41,
-	QUESTSTATEFLAG_A3INTRO = 41,
-	QUESTSTATEFLAG_A5INTRO = 41,
-};
-
-enum D2Quests
-{
-	QUEST_A1Q0_WARRIVGOSSIP = 0,
-	QUEST_A1Q1_DENOFEVIL = 1,
-	QUEST_A1Q2_BLOODRAVEN = 2,
-	QUEST_A1Q3_MALUS = 3,
-	QUEST_A1Q4_CAIN = 4,
-	QUEST_A1Q5_COUNTESS = 5,
-	QUEST_A1Q6_ANDARIEL = 6,
-
-	QUEST_A2Q0_JERHYNGOSSIP = 7,
-	QUEST_A2Q1_RADAMENT = 8,
-	QUEST_A2Q2_HORADRICSTAFF = 9,
-	QUEST_A2Q3_TAINTEDSUN = 10,
-	QUEST_A2Q4_HORAZONTOME = 11,
-	QUEST_A2Q5_SUMMONER = 12,
-	QUEST_A2Q6_DURIEL = 13,
-
-	QUEST_A3Q0_HRATLIGOSSIP = 14,
-	QUEST_A3Q1_LAMESENTOME = 15,
-	QUEST_A3Q2_KHALIMFLAIL = 16,
-	QUEST_A3Q3_GIDBINN = 17,
-	QUEST_A3Q4_GOLDENBIRD = 18,
-	QUEST_A3Q5_TRAVINCAL = 19,
-	QUEST_A3Q6_MEPHISTO = 20,
-
-	QUEST_A4Q0_TYRAELGOSSIP = 21,
-	QUEST_A4Q1_IZUAL = 22,
-	QUEST_A4Q2_DIABLO = 23,
-	QUEST_A4Q3_HELLFORGE = 24,
-
-	QUEST_A1Q1EX_NAVI = 25,
-
-	QUEST_A2Q7_UNUSED_GUARDGOSSIP = 26,
-	QUEST_A2Q8_UNUSED_GUARDGOSSIP = 27,
-
-	QUEST_A3Q7_DARKWANDERER = 28,
-
-	QUEST_A4Q4_MALACHAI = 29,
-
-	//TODO: 30 unused?
-
-	QUEST_A5Q1_SHENK = 31,
-	QUEST_A5Q2_RESCUESOLDIERS = 32,
-	QUEST_A5Q3_PRISONOFICE = 33,
-	QUEST_A5Q4_NIHLATHAK = 34,
-	QUEST_A5Q5_ANCIENTS = 35,
-	QUEST_A5Q6_BAAL = 36
-};
 
 //enum D2QuestFilterEvents
 //{
@@ -2398,451 +2286,3 @@ enum D2QuestAlert
 	QALERT_BLOCKED = 0x20
 };
 #endif // CONSTANTS_QUESTS
-
-#ifndef CONSTANTS_TXTFLAGS //Pseudo-Macro to group all constants
-enum D2C_MonStatsTxtFlagIndices
-{
-	MONSTATSFLAGINDEX_ISSPAWN = 0,
-	MONSTATSFLAGINDEX_ISMELEE,
-	MONSTATSFLAGINDEX_NORATIO,
-	MONSTATSFLAGINDEX_OPENDOORS,
-	MONSTATSFLAGINDEX_SETBOSS,
-	MONSTATSFLAGINDEX_BOSSXFER,
-	MONSTATSFLAGINDEX_BOSS,
-	MONSTATSFLAGINDEX_PRIMEEVIL,
-	MONSTATSFLAGINDEX_NPC,
-	MONSTATSFLAGINDEX_INTERACT,
-	MONSTATSFLAGINDEX_INTOWN,
-	MONSTATSFLAGINDEX_LUNDEAD,
-	MONSTATSFLAGINDEX_HUNDEAD,
-	MONSTATSFLAGINDEX_DEMON,
-	MONSTATSFLAGINDEX_FLYING,
-	MONSTATSFLAGINDEX_KILLABLE,
-	MONSTATSFLAGINDEX_SWITCHAI,
-	MONSTATSFLAGINDEX_NOMULTISHOT,
-	MONSTATSFLAGINDEX_NEVERCOUNT,
-	MONSTATSFLAGINDEX_PETIGNORE,
-	MONSTATSFLAGINDEX_DEATHDMG,
-	MONSTATSFLAGINDEX_GENERICSPAWN,
-	MONSTATSFLAGINDEX_ZOO,
-	MONSTATSFLAGINDEX_PLACESPAWN,
-	MONSTATSFLAGINDEX_INVENTORY,
-	MONSTATSFLAGINDEX_ENABLED,
-	MONSTATSFLAGINDEX_NOSHLDBLOCK,
-	MONSTATSFLAGINDEX_NOAURA,
-	MONSTATSFLAGINDEX_RANGEDTYPE,
-};
-
-enum D2C_MonStatsTxtFlags
-{
-	MONSTATSFLAG_ISSPAWN = (1 << MONSTATSFLAGINDEX_ISSPAWN),
-	MONSTATSFLAG_ISMELEE = (1 << MONSTATSFLAGINDEX_ISMELEE),
-	MONSTATSFLAG_NORATIO = (1 << MONSTATSFLAGINDEX_NORATIO),
-	MONSTATSFLAG_OPENDOORS = (1 << MONSTATSFLAGINDEX_OPENDOORS),
-	MONSTATSFLAG_SETBOSS = (1 << MONSTATSFLAGINDEX_SETBOSS),
-	MONSTATSFLAG_BOSSXFER = (1 << MONSTATSFLAGINDEX_BOSSXFER),
-	MONSTATSFLAG_BOSS = (1 << MONSTATSFLAGINDEX_BOSS),
-	MONSTATSFLAG_PRIMEEVIL = (1 << MONSTATSFLAGINDEX_PRIMEEVIL),
-	MONSTATSFLAG_NPC = (1 << MONSTATSFLAGINDEX_NPC),
-	MONSTATSFLAG_INTERACT = (1 << MONSTATSFLAGINDEX_INTERACT),
-	MONSTATSFLAG_INTOWN = (1 << MONSTATSFLAGINDEX_INTOWN),
-	MONSTATSFLAG_LUNDEAD = (1 << MONSTATSFLAGINDEX_LUNDEAD),
-	MONSTATSFLAG_HUNDEAD = (1 << MONSTATSFLAGINDEX_HUNDEAD),
-	MONSTATSFLAG_DEMON = (1 << MONSTATSFLAGINDEX_DEMON),
-	MONSTATSFLAG_FLYING = (1 << MONSTATSFLAGINDEX_FLYING),
-	MONSTATSFLAG_KILLABLE = (1 << MONSTATSFLAGINDEX_KILLABLE),
-	MONSTATSFLAG_SWITCHAI = (1 << MONSTATSFLAGINDEX_SWITCHAI),
-	MONSTATSFLAG_NOMULTISHOT = (1 << MONSTATSFLAGINDEX_NOMULTISHOT),
-	MONSTATSFLAG_NEVERCOUNT = (1 << MONSTATSFLAGINDEX_NEVERCOUNT),
-	MONSTATSFLAG_PETIGNORE = (1 << MONSTATSFLAGINDEX_PETIGNORE),
-	MONSTATSFLAG_DEATHDMG = (1 << MONSTATSFLAGINDEX_DEATHDMG),
-	MONSTATSFLAG_GENERICSPAWN = (1 << MONSTATSFLAGINDEX_GENERICSPAWN),
-	MONSTATSFLAG_ZOO = (1 << MONSTATSFLAGINDEX_ZOO),
-	MONSTATSFLAG_PLACESPAWN = (1 << MONSTATSFLAGINDEX_PLACESPAWN),
-	MONSTATSFLAG_INVENTORY = (1 << MONSTATSFLAGINDEX_INVENTORY),
-	MONSTATSFLAG_ENABLED = (1 << MONSTATSFLAGINDEX_ENABLED),
-	MONSTATSFLAG_NOSHLDBLOCK = (1 << MONSTATSFLAGINDEX_NOSHLDBLOCK),
-	MONSTATSFLAG_NOAURA = (1 << MONSTATSFLAGINDEX_NOAURA),
-	MONSTATSFLAG_RANGEDTYPE = (1 << MONSTATSFLAGINDEX_RANGEDTYPE),
-};
-
-enum D2C_MonStats2TxtFlagIndices
-{
-	MONSTATS2FLAGINDEX_NOGFXHITTEST = 0,
-	MONSTATS2FLAGINDEX_NOMAP,
-	MONSTATS2FLAGINDEX_NOOVLY,
-	MONSTATS2FLAGINDEX_ISSEL,
-	MONSTATS2FLAGINDEX_ALSEL,
-	MONSTATS2FLAGINDEX_NOSEL,
-	MONSTATS2FLAGINDEX_SHIFTSEL,
-	MONSTATS2FLAGINDEX_CORPSESEL,
-	MONSTATS2FLAGINDEX_REVIVE,
-	MONSTATS2FLAGINDEX_ISATT,
-	MONSTATS2FLAGINDEX_SMALL,
-	MONSTATS2FLAGINDEX_LARGE,
-	MONSTATS2FLAGINDEX_SOFT,
-	MONSTATS2FLAGINDEX_CRITTER,
-	MONSTATS2FLAGINDEX_SHADOW,
-	MONSTATS2FLAGINDEX_NOUNIQUESHIFT,
-	MONSTATS2FLAGINDEX_COMPOSITEDEATH,
-	MONSTATS2FLAGINDEX_INERT,
-	MONSTATS2FLAGINDEX_OBJCOL,
-	MONSTATS2FLAGINDEX_DEADCOL,
-	MONSTATS2FLAGINDEX_UNFLATDEAD,
-};
-
-enum D2C_MonStats2TxtFlags
-{
-	MONSTATS2FLAG_NOGFXHITTEST = (1 << MONSTATS2FLAGINDEX_NOGFXHITTEST),
-	MONSTATS2FLAG_NOMAP = (1 << MONSTATS2FLAGINDEX_NOMAP),
-	MONSTATS2FLAG_NOOVLY = (1 << MONSTATS2FLAGINDEX_NOOVLY),
-	MONSTATS2FLAG_ISSEL = (1 << MONSTATS2FLAGINDEX_ISSEL),
-	MONSTATS2FLAG_ALSEL = (1 << MONSTATS2FLAGINDEX_ALSEL),
-	MONSTATS2FLAG_NOSEL = (1 << MONSTATS2FLAGINDEX_NOSEL),
-	MONSTATS2FLAG_SHIFTSEL = (1 << MONSTATS2FLAGINDEX_SHIFTSEL),
-	MONSTATS2FLAG_CORPSESEL = (1 << MONSTATS2FLAGINDEX_CORPSESEL),
-	MONSTATS2FLAG_REVIVE = (1 << MONSTATS2FLAGINDEX_REVIVE),
-	MONSTATS2FLAG_ISATT = (1 << MONSTATS2FLAGINDEX_ISATT),
-	MONSTATS2FLAG_SMALL = (1 << MONSTATS2FLAGINDEX_SMALL),
-	MONSTATS2FLAG_LARGE = (1 << MONSTATS2FLAGINDEX_LARGE),
-	MONSTATS2FLAG_SOFT = (1 << MONSTATS2FLAGINDEX_SOFT),
-	MONSTATS2FLAG_CRITTER = (1 << MONSTATS2FLAGINDEX_CRITTER),
-	MONSTATS2FLAG_SHADOW = (1 << MONSTATS2FLAGINDEX_SHADOW),
-	MONSTATS2FLAG_NOUNIQUESHIFT = (1 << MONSTATS2FLAGINDEX_NOUNIQUESHIFT),
-	MONSTATS2FLAG_COMPOSITEDEATH = (1 << MONSTATS2FLAGINDEX_COMPOSITEDEATH),
-	MONSTATS2FLAG_INERT = (1 << MONSTATS2FLAGINDEX_INERT),
-	MONSTATS2FLAG_OBJCOL = (1 << MONSTATS2FLAGINDEX_OBJCOL),
-	MONSTATS2FLAG_DEADCOL = (1 << MONSTATS2FLAGINDEX_DEADCOL),
-	MONSTATS2FLAG_UNFLATDEAD = (1 << MONSTATS2FLAGINDEX_UNFLATDEAD),
-};
-
-enum D2C_MonStats2TxtComponentFlagIndices
-{
-	MONSTATS2COMPONENTFLAGINDEX_HD = 0,
-	MONSTATS2COMPONENTFLAGINDEX_TR,
-	MONSTATS2COMPONENTFLAGINDEX_LG,
-	MONSTATS2COMPONENTFLAGINDEX_RA,
-	MONSTATS2COMPONENTFLAGINDEX_LA,
-	MONSTATS2COMPONENTFLAGINDEX_RH,
-	MONSTATS2COMPONENTFLAGINDEX_LH,
-	MONSTATS2COMPONENTFLAGINDEX_SH,
-	MONSTATS2COMPONENTFLAGINDEX_S1,
-	MONSTATS2COMPONENTFLAGINDEX_S2,
-	MONSTATS2COMPONENTFLAGINDEX_S3,
-	MONSTATS2COMPONENTFLAGINDEX_S4,
-	MONSTATS2COMPONENTFLAGINDEX_S5,
-	MONSTATS2COMPONENTFLAGINDEX_S6,
-	MONSTATS2COMPONENTFLAGINDEX_S7,
-	MONSTATS2COMPONENTFLAGINDEX_S8,
-};
-
-enum D2C_MonStats2TxtComponentFlags
-{
-	MONSTATS2COMPONENTFLAG_HD = (1 << MONSTATS2COMPONENTFLAGINDEX_HD),
-	MONSTATS2COMPONENTFLAG_TR = (1 << MONSTATS2COMPONENTFLAGINDEX_TR),
-	MONSTATS2COMPONENTFLAG_LG = (1 << MONSTATS2COMPONENTFLAGINDEX_LG),
-	MONSTATS2COMPONENTFLAG_RA = (1 << MONSTATS2COMPONENTFLAGINDEX_RA),
-	MONSTATS2COMPONENTFLAG_LA = (1 << MONSTATS2COMPONENTFLAGINDEX_LA),
-	MONSTATS2COMPONENTFLAG_RH = (1 << MONSTATS2COMPONENTFLAGINDEX_RH),
-	MONSTATS2COMPONENTFLAG_LH = (1 << MONSTATS2COMPONENTFLAGINDEX_LH),
-	MONSTATS2COMPONENTFLAG_SH = (1 << MONSTATS2COMPONENTFLAGINDEX_SH),
-	MONSTATS2COMPONENTFLAG_S1 = (1 << MONSTATS2COMPONENTFLAGINDEX_S1),
-	MONSTATS2COMPONENTFLAG_S2 = (1 << MONSTATS2COMPONENTFLAGINDEX_S2),
-	MONSTATS2COMPONENTFLAG_S3 = (1 << MONSTATS2COMPONENTFLAGINDEX_S3),
-	MONSTATS2COMPONENTFLAG_S4 = (1 << MONSTATS2COMPONENTFLAGINDEX_S4),
-	MONSTATS2COMPONENTFLAG_S5 = (1 << MONSTATS2COMPONENTFLAGINDEX_S5),
-	MONSTATS2COMPONENTFLAG_S6 = (1 << MONSTATS2COMPONENTFLAGINDEX_S6),
-	MONSTATS2COMPONENTFLAG_S7 = (1 << MONSTATS2COMPONENTFLAGINDEX_S7),
-	MONSTATS2COMPONENTFLAG_S8 = (1 << MONSTATS2COMPONENTFLAGINDEX_S8),
-};
-
-enum D2C_MonStats2TxtModeFlagIndices
-{
-	MONSTATS2MODEFLAGINDEX_DT = 0,
-	MONSTATS2MODEFLAGINDEX_NU,
-	MONSTATS2MODEFLAGINDEX_WL,
-	MONSTATS2MODEFLAGINDEX_GH,
-	MONSTATS2MODEFLAGINDEX_A1,
-	MONSTATS2MODEFLAGINDEX_A2,
-	MONSTATS2MODEFLAGINDEX_BL,
-	MONSTATS2MODEFLAGINDEX_SC,
-	MONSTATS2MODEFLAGINDEX_S1,
-	MONSTATS2MODEFLAGINDEX_S2,
-	MONSTATS2MODEFLAGINDEX_S3,
-	MONSTATS2MODEFLAGINDEX_S4,
-	MONSTATS2MODEFLAGINDEX_DD,
-	MONSTATS2MODEFLAGINDEX_KB,
-	MONSTATS2MODEFLAGINDEX_SQ,
-	MONSTATS2MODEFLAGINDEX_RN,
-};
-
-enum D2C_MonStats2TxtModeFlags
-{
-	MONSTATS2MODEFLAG_DT = (1 << MONSTATS2MODEFLAGINDEX_DT),
-	MONSTATS2MODEFLAG_NU = (1 << MONSTATS2MODEFLAGINDEX_NU),
-	MONSTATS2MODEFLAG_WL = (1 << MONSTATS2MODEFLAGINDEX_WL),
-	MONSTATS2MODEFLAG_GH = (1 << MONSTATS2MODEFLAGINDEX_GH),
-	MONSTATS2MODEFLAG_A1 = (1 << MONSTATS2MODEFLAGINDEX_A1),
-	MONSTATS2MODEFLAG_A2 = (1 << MONSTATS2MODEFLAGINDEX_A2),
-	MONSTATS2MODEFLAG_BL = (1 << MONSTATS2MODEFLAGINDEX_BL),
-	MONSTATS2MODEFLAG_SC = (1 << MONSTATS2MODEFLAGINDEX_SC),
-	MONSTATS2MODEFLAG_S1 = (1 << MONSTATS2MODEFLAGINDEX_S1),
-	MONSTATS2MODEFLAG_S2 = (1 << MONSTATS2MODEFLAGINDEX_S2),
-	MONSTATS2MODEFLAG_S3 = (1 << MONSTATS2MODEFLAGINDEX_S3),
-	MONSTATS2MODEFLAG_S4 = (1 << MONSTATS2MODEFLAGINDEX_S4),
-	MONSTATS2MODEFLAG_DD = (1 << MONSTATS2MODEFLAGINDEX_DD),
-	MONSTATS2MODEFLAG_KB = (1 << MONSTATS2MODEFLAGINDEX_KB),
-	MONSTATS2MODEFLAG_SQ = (1 << MONSTATS2MODEFLAGINDEX_SQ),
-	MONSTATS2MODEFLAG_RN = (1 << MONSTATS2MODEFLAGINDEX_RN),
-};
-
-enum D2C_PetTypeTxtFlagIndices
-{
-	PETTYPEFLAGINDEX_WARP = 0,
-	PETTYPEFLAGINDEX_RANGE,
-	PETTYPEFLAGINDEX_PARTYSEND,
-	PETTYPEFLAGINDEX_UNSUMMON,
-	PETTYPEFLAGINDEX_AUTOMAP,
-	PETTYPEFLAGINDEX_DRAWHP,
-};
-
-enum D2C_PetTypeTxtFlags
-{
-	PETTYPEFLAG_WARP = (1 << PETTYPEFLAGINDEX_WARP),
-	PETTYPEFLAG_RANGE = (1 << PETTYPEFLAGINDEX_RANGE),
-	PETTYPEFLAG_PARTYSEND = (1 << PETTYPEFLAGINDEX_PARTYSEND),
-	PETTYPEFLAG_UNSUMMON = (1 << PETTYPEFLAGINDEX_UNSUMMON),
-	PETTYPEFLAG_AUTOMAP = (1 << PETTYPEFLAGINDEX_AUTOMAP),
-	PETTYPEFLAG_DRAWHP = (1 << PETTYPEFLAGINDEX_DRAWHP),
-};
-
-enum D2C_ItemStatCostTxtFlagIndices
-{
-	ITEMSTATCOSTFLAGINDEX_SEND_OTHER = 0,
-	ITEMSTATCOSTFLAGINDEX_SIGNED,
-	ITEMSTATCOSTFLAGINDEX_DAMAGERELATED,
-	ITEMSTATCOSTFLAGINDEX_ITEMSPECIFIC,
-	ITEMSTATCOSTFLAGINDEX_DIRECT,
-
-	ITEMSTATCOSTFLAGINDEX_HP_MANA_STAMINA = 5,
-	ITEMSTATCOSTFLAGINDEX_HP,
-	ITEMSTATCOSTFLAGINDEX_MANA,
-	ITEMSTATCOSTFLAGINDEX_STAMINA,
-
-	ITEMSTATCOSTFLAGINDEX_UPDATEANIMRATE = 9,
-	ITEMSTATCOSTFLAGINDEX_FMIN,
-	ITEMSTATCOSTFLAGINDEX_FCALLBACK,
-	ITEMSTATCOSTFLAGINDEX_SAVED,
-	ITEMSTATCOSTFLAGINDEX_CSVSIGNED,
-};
-
-enum D2C_ItemStatCostTxtFlags
-{
-	ITEMSTATCOSTFLAG_SEND_OTHER = (1 << ITEMSTATCOSTFLAGINDEX_SEND_OTHER),
-	ITEMSTATCOSTFLAG_SIGNED = (1 << ITEMSTATCOSTFLAGINDEX_SIGNED),
-	ITEMSTATCOSTFLAG_DAMAGERELATED = (1 << ITEMSTATCOSTFLAGINDEX_DAMAGERELATED),
-	ITEMSTATCOSTFLAG_ITEMSPECIFIC = (1 << ITEMSTATCOSTFLAGINDEX_ITEMSPECIFIC),
-	ITEMSTATCOSTFLAG_DIRECT = (1 << ITEMSTATCOSTFLAGINDEX_DIRECT),
-
-	ITEMSTATCOSTFLAG_HP_MANA_STAMINA = (1 << ITEMSTATCOSTFLAGINDEX_HP_MANA_STAMINA),
-	ITEMSTATCOSTFLAG_HP = (1 << ITEMSTATCOSTFLAGINDEX_HP),
-	ITEMSTATCOSTFLAG_MANA = (1 << ITEMSTATCOSTFLAGINDEX_MANA),
-	ITEMSTATCOSTFLAG_STAMINA = (1 << ITEMSTATCOSTFLAGINDEX_STAMINA),
-
-	ITEMSTATCOSTFLAG_UPDATEANIMRATE = (1 << ITEMSTATCOSTFLAGINDEX_UPDATEANIMRATE),
-	ITEMSTATCOSTFLAG_FMIN = (1 << ITEMSTATCOSTFLAGINDEX_FMIN),
-	ITEMSTATCOSTFLAG_FCALLBACK = (1 << ITEMSTATCOSTFLAGINDEX_FCALLBACK),
-	ITEMSTATCOSTFLAG_SAVED = (1 << ITEMSTATCOSTFLAGINDEX_SAVED),
-	ITEMSTATCOSTFLAG_CSVSIGNED = (1 << ITEMSTATCOSTFLAGINDEX_CSVSIGNED),
-};
-
-enum D2C_UniqueItemsTxtFlagIndices
-{
-	UNIQUEITEMSFLAGINDEX_ENABLED = 0,
-	UNIQUEITEMSFLAGINDEX_NOLIMIT,
-	UNIQUEITEMSFLAGINDEX_CARRY1,
-	UNIQUEITEMSFLAGINDEX_LADDER,
-};
-
-enum D2C_UniqueItemsTxtFlags
-{
-	UNIQUEITEMSFLAG_ENABLED = (1 << UNIQUEITEMSFLAGINDEX_ENABLED),
-	UNIQUEITEMSFLAG_NOLIMIT = (1 << UNIQUEITEMSFLAGINDEX_NOLIMIT),
-	UNIQUEITEMSFLAG_CARRY1 = (1 << UNIQUEITEMSFLAGINDEX_CARRY1),
-	UNIQUEITEMSFLAG_LADDER = (1 << UNIQUEITEMSFLAGINDEX_LADDER),
-};
-
-enum D2C_MissilesTxtFlagIndices
-{
-	MISSILESFLAGINDEX_LASTCOLLIDE = 0,
-	MISSILESFLAGINDEX_EXPLOSION,
-	MISSILESFLAGINDEX_PIERCE,
-	MISSILESFLAGINDEX_CANSLOW,
-	MISSILESFLAGINDEX_CANDESTROY,
-	MISSILESFLAGINDEX_CLIENTSEND,
-	MISSILESFLAGINDEX_GETHIT,
-	MISSILESFLAGINDEX_SOFTHIT,
-	MISSILESFLAGINDEX_APPLYMASTERY,
-	MISSILESFLAGINDEX_RETURNFIRE,
-	MISSILESFLAGINDEX_TOWN,
-	MISSILESFLAGINDEX_SRCTOWN,
-	MISSILESFLAGINDEX_NOMULTISHOT,
-	MISSILESFLAGINDEX_NOUNIQUEMOD,
-	MISSILESFLAGINDEX_HALF2HSRC,
-	MISSILESFLAGINDEX_MISSILESKILL,
-};
-
-enum D2C_MissilesTxtFlags
-{
-	MISSILESFLAG_LASTCOLLIDE = (1 << MISSILESFLAGINDEX_LASTCOLLIDE),
-	MISSILESFLAG_EXPLOSION = (1 << MISSILESFLAGINDEX_EXPLOSION),
-	MISSILESFLAG_PIERCE = (1 << MISSILESFLAGINDEX_PIERCE),
-	MISSILESFLAG_CANSLOW = (1 << MISSILESFLAGINDEX_CANSLOW),
-	MISSILESFLAG_CANDESTROY = (1 << MISSILESFLAGINDEX_CANDESTROY),
-	MISSILESFLAG_CLIENTSEND = (1 << MISSILESFLAGINDEX_CLIENTSEND),
-	MISSILESFLAG_GETHIT = (1 << MISSILESFLAGINDEX_GETHIT),
-	MISSILESFLAG_SOFTHIT = (1 << MISSILESFLAGINDEX_SOFTHIT),
-	MISSILESFLAG_APPLYMASTERY = (1 << MISSILESFLAGINDEX_APPLYMASTERY),
-	MISSILESFLAG_RETURNFIRE = (1 << MISSILESFLAGINDEX_RETURNFIRE),
-	MISSILESFLAG_TOWN = (1 << MISSILESFLAGINDEX_TOWN),
-	MISSILESFLAG_SRCTOWN = (1 << MISSILESFLAGINDEX_SRCTOWN),
-	MISSILESFLAG_NOMULTISHOT = (1 << MISSILESFLAGINDEX_NOMULTISHOT),
-	MISSILESFLAG_NOUNIQUEMOD = (1 << MISSILESFLAGINDEX_NOUNIQUEMOD),
-	MISSILESFLAG_HALF2HSRC = (1 << MISSILESFLAGINDEX_HALF2HSRC),
-	MISSILESFLAG_MISSILESKILL = (1 << MISSILESFLAGINDEX_MISSILESKILL),
-};
-
-enum D2C_SkillsTxtFlagIndices
-{
-	SKILLSFLAGINDEX_DECQUANT = 0,
-	SKILLSFLAGINDEX_LOB,
-	SKILLSFLAGINDEX_PROGRESSIVE,
-	SKILLSFLAGINDEX_FINISHING,
-	SKILLSFLAGINDEX_PASSIVE,
-	SKILLSFLAGINDEX_AURA,
-	SKILLSFLAGINDEX_PERIODIC,
-	SKILLSFLAGINDEX_PRGSTACK,
-	SKILLSFLAGINDEX_INTOWN,
-	SKILLSFLAGINDEX_KICK,
-	SKILLSFLAGINDEX_INGAME,
-	SKILLSFLAGINDEX_REPEAT,
-	SKILLSFLAGINDEX_STSUCCESSONLY,
-	SKILLSFLAGINDEX_STSOUNDDELAY,
-	SKILLSFLAGINDEX_WEAPONSND,
-	SKILLSFLAGINDEX_IMMEDIATE,
-	SKILLSFLAGINDEX_NOAMMO,
-	SKILLSFLAGINDEX_ENHANCEABLE,
-	SKILLSFLAGINDEX_DURABILITY,
-	SKILLSFLAGINDEX_USEATTACKRATE,
-	SKILLSFLAGINDEX_TARGETABLEONLY,
-	SKILLSFLAGINDEX_SEARCHENEMYXY,
-	SKILLSFLAGINDEX_SEARCHENEMYNEAR,
-	SKILLSFLAGINDEX_SEARCHOPENXY,
-	SKILLSFLAGINDEX_TARGETCORPSE,
-	SKILLSFLAGINDEX_TARGETPET,
-	SKILLSFLAGINDEX_TARGETALLY,
-	SKILLSFLAGINDEX_TARGETITEM,
-	SKILLSFLAGINDEX_ATTACKNOMANA,
-	SKILLSFLAGINDEX_ITEMTGTDO,
-	SKILLSFLAGINDEX_LEFTSKILL,
-	SKILLSFLAGINDEX_INTERRUPT,
-
-	//Bits above 31 need to restart at 0
-	SKILLSFLAGINDEX2_TGTPLACECHECK = 0,
-	SKILLSFLAGINDEX2_ITEMCHECKSTART,
-	SKILLSFLAGINDEX2_ITEMCLTCHECKSTART,
-	SKILLSFLAGINDEX2_GENERAL,
-	SKILLSFLAGINDEX2_SCROLL,
-	SKILLSFLAGINDEX2_USEMANAONDO,
-	SKILLSFLAGINDEX2_WARP,
-};
-
-enum D2C_SkillsTxtFlags
-{
-	SKILLSFLAG_DECQUANT = (1 << SKILLSFLAGINDEX_DECQUANT),
-	SKILLSFLAG_LOB = (1 << SKILLSFLAGINDEX_LOB),
-	SKILLSFLAG_PROGRESSIVE = (1 << SKILLSFLAGINDEX_PROGRESSIVE),
-	SKILLSFLAG_FINISHING = (1 << SKILLSFLAGINDEX_FINISHING),
-	SKILLSFLAG_PASSIVE = (1 << SKILLSFLAGINDEX_PASSIVE),
-	SKILLSFLAG_AURA = (1 << SKILLSFLAGINDEX_AURA),
-	SKILLSFLAG_PERIODIC = (1 << SKILLSFLAGINDEX_PERIODIC),
-	SKILLSFLAG_PRGSTACK = (1 << SKILLSFLAGINDEX_PRGSTACK),
-	SKILLSFLAG_INTOWN = (1 << SKILLSFLAGINDEX_INTOWN),
-	SKILLSFLAG_KICK = (1 << SKILLSFLAGINDEX_KICK),
-	SKILLSFLAG_INGAME = (1 << SKILLSFLAGINDEX_INGAME),
-	SKILLSFLAG_REPEAT = (1 << SKILLSFLAGINDEX_REPEAT),
-	SKILLSFLAG_STSUCCESSONLY = (1 << SKILLSFLAGINDEX_STSUCCESSONLY),
-	SKILLSFLAG_STSOUNDDELAY = (1 << SKILLSFLAGINDEX_STSOUNDDELAY),
-	SKILLSFLAG_WEAPONSND = (1 << SKILLSFLAGINDEX_WEAPONSND),
-	SKILLSFLAG_IMMEDIATE = (1 << SKILLSFLAGINDEX_IMMEDIATE),
-	SKILLSFLAG_NOAMMO = (1 << SKILLSFLAGINDEX_NOAMMO),
-	SKILLSFLAG_ENHANCEABLE = (1 << SKILLSFLAGINDEX_ENHANCEABLE),
-	SKILLSFLAG_DURABILITY = (1 << SKILLSFLAGINDEX_DURABILITY),
-	SKILLSFLAG_USEATTACKRATE = (1 << SKILLSFLAGINDEX_USEATTACKRATE),
-	SKILLSFLAG_TARGETABLEONLY = (1 << SKILLSFLAGINDEX_TARGETABLEONLY),
-	SKILLSFLAG_SEARCHENEMYXY = (1 << SKILLSFLAGINDEX_SEARCHENEMYXY),
-	SKILLSFLAG_SEARCHENEMYNEAR = (1 << SKILLSFLAGINDEX_SEARCHENEMYNEAR),
-	SKILLSFLAG_SEARCHOPENXY = (1 << SKILLSFLAGINDEX_SEARCHOPENXY),
-	SKILLSFLAG_TARGETCORPSE = (1 << SKILLSFLAGINDEX_TARGETCORPSE),
-	SKILLSFLAG_TARGETPET = (1 << SKILLSFLAGINDEX_TARGETPET),
-	SKILLSFLAG_TARGETALLY = (1 << SKILLSFLAGINDEX_TARGETALLY),
-	SKILLSFLAG_TARGETITEM = (1 << SKILLSFLAGINDEX_TARGETITEM),
-	SKILLSFLAG_ATTACKNOMANA = (1 << SKILLSFLAGINDEX_ATTACKNOMANA),
-	SKILLSFLAG_ITEMTGTDO = (1 << SKILLSFLAGINDEX_ITEMTGTDO),
-	SKILLSFLAG_LEFTSKILL = (1 << SKILLSFLAGINDEX_LEFTSKILL),
-	SKILLSFLAG_INTERRUPT = (1 << SKILLSFLAGINDEX_INTERRUPT),
-
-	//Bits above 31 need to restart at 0
-	SKILLSFLAG2_TGTPLACECHECK = (1 << SKILLSFLAGINDEX2_TGTPLACECHECK),
-	SKILLSFLAG2_ITEMCHECKSTART = (1 << SKILLSFLAGINDEX2_ITEMCHECKSTART),
-	SKILLSFLAG2_ITEMCLTCHECKSTART = (1 << SKILLSFLAGINDEX2_ITEMCLTCHECKSTART),
-	SKILLSFLAG2_GENERAL = (1 << SKILLSFLAGINDEX2_GENERAL),
-	SKILLSFLAG2_SCROLL = (1 << SKILLSFLAGINDEX2_SCROLL),
-	SKILLSFLAG2_USEMANAONDO = (1 << SKILLSFLAGINDEX2_USEMANAONDO),
-	SKILLSFLAG2_WARP = (1 << SKILLSFLAGINDEX2_WARP),
-};
-
-enum D2C_StateMasks
-{
-	STATEMASK_NOSEND,
-	STATEMASK_AURA,
-	STATEMASK_HIDE,
-	STATEMASK_TRANSFORM,
-	STATEMASK_PGSV,
-	STATEMASK_ACTIVE,
-	STATEMASK_REMHIT,
-	STATEMASK_DAMBLUE,
-	STATEMASK_DAMRED,
-	STATEMASK_ATTBLUE,
-	STATEMASK_ATTRED,
-	STATEMASK_CURSE,
-	STATEMASK_CURABLE,
-	STATEMASK_PLRSTAYDEATH,
-	STATEMASK_MONSTAYDEATH,
-	STATEMASK_BOSSSTAYDEATH,
-	STATEMASK_DISGUISE,
-	STATEMASK_RESTRICT,
-	STATEMASK_BLUE,
-	STATEMASK_ARMBLUE,
-	STATEMASK_RFBLUE,
-	STATEMASK_RCBLUE,
-	STATEMASK_RLBLUE,
-	STATEMASK_RPBLUE,
-	STATEMASK_STAMBARBLUE,
-	STATEMASK_ARMRED,
-	STATEMASK_RFRED,
-	STATEMASK_RCRED,
-	STATEMASK_RLRED,
-	STATEMASK_RPRED,
-	STATEMASK_EXP,
-	STATEMASK_SHATTER,
-	STATEMASK_LIFE,
-	STATEMASK_UDEAD,
-	STATEMASK_GREEN,
-	STATEMASK_NOOVERLAYS,
-	STATEMASK_NOCLEAR,
-	STATEMASK_BOSSINV,
-	STATEMASK_MELEEONLY,
-	STATEMASK_NOTONDEAD,
-};
-#endif // CONSTANTS_TXTFLAGS
