+++ source/D2CommonDefinitions/include/D2QuestDataEx.h
@@ -13,6 +13,118 @@ struct D2CoordStrc							//sizeof 0x08
 	int nY;										//0x04
 };
 
+enum D2QuestStateFlagIds
+{
+	QUESTSTATEFLAG_A1Q0 = 0,
+	QUESTSTATEFLAG_A1Q1 = 1,
+	QUESTSTATEFLAG_A1Q2 = 2,
+	QUESTSTATEFLAG_A1Q3 = 3,
+	QUESTSTATEFLAG_A1Q4 = 4,
+	QUESTSTATEFLAG_A1Q5 = 5,
+	QUESTSTATEFLAG_A1Q6 = 6,
+
+	QUESTSTATEFLAG_A1COMPLETED = 7,
+
+	QUESTSTATEFLAG_A2Q0 = 8,
+	QUESTSTATEFLAG_A2Q1 = 9,
+	QUESTSTATEFLAG_A2Q2 = 10,
+	QUESTSTATEFLAG_A2Q3 = 11,
+	QUESTSTATEFLAG_A2Q4 = 12,
+	QUESTSTATEFLAG_A2Q5 = 13,
+	QUESTSTATEFLAG_A2Q6 = 14,
+
+	QUESTSTATEFLAG_A2COMPLETED = 15,
+
+	QUESTSTATEFLAG_A3Q0 = 16,
+	QUESTSTATEFLAG_A3Q1 = 17,
+	QUESTSTATEFLAG_A3Q2 = 18,
+	QUESTSTATEFLAG_A3Q3 = 19,
+	QUESTSTATEFLAG_A3Q4 = 20,
+	QUESTSTATEFLAG_A3Q5 = 21,
+	QUESTSTATEFLAG_A3Q6 = 22,
+
+	QUESTSTATEFLAG_A3COMPLETED = 23,
+
+	QUESTSTATEFLAG_A4Q0 = 24,
+	QUESTSTATEFLAG_A4Q1 = 25,
+	QUESTSTATEFLAG_A4Q2 = 26,
+	QUESTSTATEFLAG_A4Q3 = 27,
+
+	QUESTSTATEFLAG_A4COMPLETED = 28,
+
+	QUESTSTATEFLAG_A1Q7 = 29,
+
+	QUESTSTATEFLAG_A2Q7 = 30,
+	QUESTSTATEFLAG_A2Q8 = 31,
+
+	QUESTSTATEFLAG_A3Q7 = 32,
+
+	QUESTSTATEFLAG_A4Q4 = 33,
+
+	//Unused: 34
+
+	QUESTSTATEFLAG_A5Q1 = 35,
+	QUESTSTATEFLAG_A5Q2 = 36,
+	QUESTSTATEFLAG_A5Q3 = 37,
+	QUESTSTATEFLAG_A5Q4 = 38,
+	QUESTSTATEFLAG_A5Q5 = 39,
+	QUESTSTATEFLAG_A5Q6 = 40,
+
+	QUESTSTATEFLAG_A1INTRO = 41,
+	QUESTSTATEFLAG_A2INTRO = 41,
+	QUESTSTATEFLAG_A3INTRO = 41,
+	QUESTSTATEFLAG_A5INTRO = 41,
+};
+
+enum D2Quests
+{
+	QUEST_A1Q0_WARRIVGOSSIP = 0,
+	QUEST_A1Q1_DENOFEVIL = 1,
+	QUEST_A1Q2_BLOODRAVEN = 2,
+	QUEST_A1Q3_MALUS = 3,
+	QUEST_A1Q4_CAIN = 4,
+	QUEST_A1Q5_COUNTESS = 5,
+	QUEST_A1Q6_ANDARIEL = 6,
+
+	QUEST_A2Q0_JERHYNGOSSIP = 7,
+	QUEST_A2Q1_RADAMENT = 8,
+	QUEST_A2Q2_HORADRICSTAFF = 9,
+	QUEST_A2Q3_TAINTEDSUN = 10,
+	QUEST_A2Q4_HORAZONTOME = 11,
+	QUEST_A2Q5_SUMMONER = 12,
+	QUEST_A2Q6_DURIEL = 13,
+
+	QUEST_A3Q0_HRATLIGOSSIP = 14,
+	QUEST_A3Q1_LAMESENTOME = 15,
+	QUEST_A3Q2_KHALIMFLAIL = 16,
+	QUEST_A3Q3_GIDBINN = 17,
+	QUEST_A3Q4_GOLDENBIRD = 18,
+	QUEST_A3Q5_TRAVINCAL = 19,
+	QUEST_A3Q6_MEPHISTO = 20,
+
+	QUEST_A4Q0_TYRAELGOSSIP = 21,
+	QUEST_A4Q1_IZUAL = 22,
+	QUEST_A4Q2_DIABLO = 23,
+	QUEST_A4Q3_HELLFORGE = 24,
+
+	QUEST_A1Q1EX_NAVI = 25,
+
+	QUEST_A2Q7_UNUSED_GUARDGOSSIP = 26,
+	QUEST_A2Q8_UNUSED_GUARDGOSSIP = 27,
+
+	QUEST_A3Q7_DARKWANDERER = 28,
+
+	QUEST_A4Q4_MALACHAI = 29,
+
+	//TODO: 30 unused?
+
+	QUEST_A5Q1_SHENK = 31,
+	QUEST_A5Q2_RESCUESOLDIERS = 32,
+	QUEST_A5Q3_PRISONOFICE = 33,
+	QUEST_A5Q4_NIHLATHAK = 34,
+	QUEST_A5Q5_ANCIENTS = 35,
+	QUEST_A5Q6_BAAL = 36
+};
 struct D2QuestGUIDStrc						//sizeof 0x84
 {
 	uint32_t dwPlayerGUID[32];						//0x00
